fix: use dtype instead of deprecated torch_dtype in from_pretrained
The torch_dtype parameter was renamed to dtype in transformers v5.

+++ streamlit_app.py
@@ -199,7 +199,7 @@ def load_model() -> tuple:
     device = DEVICE if DEVICE != "auto" else detect_device()
     dtype = select_dtype(device)
     tokenizer = AutoTokenizer.from_pretrained(MODEL_ID)
-    model = AutoModelForCausalLM.from_pretrained(MODEL_ID, torch_dtype=dtype)
+    model = AutoModelForCausalLM.from_pretrained(MODEL_ID, dtype=dtype)
     model = model.to(device).eval()
     return tokenizer, model, device, dtype
